+++ CHANGELOG.md
@@ -1,10 +1,26 @@
 # Change Log
 
-## Version 1.2.X - XYZth XYZ 2018
+## Version 1.2.X - XYZth XYZ 2019
 
 ### Features
 
 * ...
+
+### Updates
+
+* ...
+
+### Bugfixes
+
+* ...
+
+---
+
+## Version 1.2.3 - 1th January 2019
+
+### Features
+
+* New Game Tank (*digitsensitive*)
 
 ### Updates
 
@@ -31,6 +47,9 @@
 * Update tint cheatsheet (*digitsensitive*)
 * Update visible cheatsheet (*digitsensitive*)
 * Add computed-size cheatsheet (*digitsensitive*)
+* Update game cheatsheet (*digitsensitive*)
+* Update config cheatsheet (*digitsensitive*)
+* Update scene cheatsheets (*digitsensitive*)
 * Add scene folder with `plugin.md`, `settings.md` and `systems.md` to cheatsheets (*digitsensitive*)
 * Add data folder with `data-manager-plugin.md` to cheatsheets (*digitsensitive*)
 * rename window.onload to window.addEventListener (*digitsensitive*)
@@ -41,6 +60,7 @@
 
 * update syntax after the upgrade to phaser 3.15.1 (*digitsensitive*)
 
+---
 
 ## Version 1.2.2 - 17th November 2018
 
@@ -63,6 +83,7 @@
 
 * Small changes in space-invaders (*digitsensitive*)
 
+---
 
 ## Version 1.2.1 (Juno) - 7th July 2018
 
@@ -85,6 +106,7 @@
 * fixes in mainMenuScene of blockade and in gameScene of snake (*digitsensitive*)
 * fix startKey Problem in mainMenuScene of flappy bird (*digitsensitive*)
 
+---
 
 ## Version 1.2.0 (Sliver) - 2th June 2018
 
@@ -108,6 +130,7 @@
 * add mode to webpack in package.json (*emedws*)
 * small change in output of the webpack.config.js file (*digitsensitive*)
 
+---
 
 ## Version 1.1.0 - 13th May 2018
 
@@ -137,6 +160,7 @@
 * fix config error in game.ts (constructor(config: GameConfig)) files after updating to official phaser.d.ts (*vladfaust*)
 * workaround for intersect in the phaser.d.ts file (*digitsensitive*)
 
+---
 
 ## Version 1.0.9 - 17th April 2018
 
@@ -146,6 +170,7 @@
 * update webpack version from 4.5.0 to 4.6.0 (*digitsensitive*)
 * update webpack-dev-server version from 3.1.2 to 3.1.3 (*digitsensitive*)
 
+---
 
 ## Version 1.0.8 - 07th April 2018
 
@@ -161,6 +186,7 @@
 * update webpack-dev-server version from 3.1.1 to 3.1.2 (*digitsensitive*)
 * update webpack-cli version from 2.0.13 to 2.0.14 (*digitsensitive*)
 
+---
 
 ## Version 1.0.7 - 01th April 2018
 
@@ -178,6 +204,7 @@
   - this.scene.add("GameScene", GameScene, false);
   - this.scene.start("GameScene");
 
+---
 
 ## Version 1.0.6 - 25th March 2018
 
@@ -195,6 +222,7 @@
 
 * phaser.d.ts: fix RectangleToRectangle problem -> missing object
 
+---
 
 ## Version 1.0.5 - 24th March 2018
 
@@ -212,6 +240,7 @@
 * add image to game config tutorial (*digitsensitive*)
 * small changes in the README.md (*digitsensitive*)
 
+---
 
 ## Version 1.0.4 - 18th March 2018
 
@@ -219,6 +248,7 @@
 
 * update phaser version from 3.2.0 to 3.2.1 (*digitsensitive*)
 
+---
 
 ## Version 1.0.3 - 11th March 2018